Remove sentence about affiliation from steering committee (#12)
I propose to remove this sentence as it may have some unintended implications. First, it sounds like we are asking organizations to commit to the policies when it's always the individuals. I think the fact that organizations have to allow individuals to bind themselves to our policies is implied already. The bigger issue with this sentence is that it talks about "because of your affiliation" which makes it sound like the position is bound to the organization, not the individual. This could the interpreted as saying that if the person leaves the organization, the organization gets to fill the slot rather than the person continuing to be on the steering committee.